Fix handling for full URLs
This fixes a bug where if the user passed a full URL, including the https:// prefix, Shorty Links would inject an extraneous https// into the link. For example:

https://whatgotdone.com/dansult/2021-10-22 -> https://https//whatgotdone.com/dansult/2021-10-22

The root cause was a brittle workaround for location.replace where it prefixed the target link with // but that only works for URLs without schemes (e.g., google.com).

The // workaround also breaks for HTTP-only sites. For example, redirecting to neverssl.com breaks because it becomes //neverssl.com, and because Shorty Links is being served over SSL, //neverssl.com becomes https://neverssl.com. But really it's irrelevant whether Shorty Links is being served over SSL. If we don't know the scheme of the target, we have to guess either HTTP or HTTPS, so this implementation guesses that it's HTTP, though we could easily change the assumption to HTTPS.